chore: Add header files
You never, *NEVER*, put a pragma once on a .c file
