+++ include/sortAlgorithms.h
@@ -1,0 +1,13 @@
+#pragma once
+
+void insertionSort(int array[], int length);
+int partition(int array[], int start, int end);
+void quickSort(int array[], int start, int end);
+
+void mergeSort(int vetor[], int inicio, int fim);
+void mergeSortRecusao(int vetor[], int inicio, int fim, int esq[], int dir[]);
+void merge(int vetor[], int inicio, int meio, int fim, int esq[], int dir[]);
+
+void printVetor(int v[], int tamanho);
+void printArrayInteval(int v[], int start, int end);
+int organized(int vetor[], int tamanho);
