+++ src/sortAlgorithms.c
@@ -1,19 +1,7 @@
-#pragma once
-
+#include "../include/sortAlgorithms.h"
 #include <stdio.h>
 #include <stdlib.h>
 #include <time.h>
-
-
-int partition(int array[], int start, int end);
-void quickSort(int array[], int start, int end);
-
-void mergeSort(int vetor[], int inicio, int fim);
-void mergeSortRecusao(int vetor[], int inicio, int fim, int esq[], int dir[]);
-void merge(int vetor[], int inicio, int meio, int fim, int esq[], int dir[]);
-
-void printVetor(int v[], int tamanho);
-void printArrayInteval(int v[], int start, int end);
 
 void insertionSort(int array[], int length)
 {
